Add files via upload
TASK 04
Task 4.1
You are selected as a data scientist in Zebtech , the manager assigned you a task to create a
program to add first ten customer details. The manager asked you to create a table that ask 3 major
question. • Customer name • Customer height • Customer contact # These questions are asked by
first ten customer. Make sure the the height should be in decimal and authentication of contact
number is mandatory , only number is allowed
Task 4.2
Design an online admission for Skillvention. The fields are same as mentioned in the admission form
Task 4.3
Create a login and registration form for your website "Python Agency